refactor: share render form logic across core and widgets
