+++ src/DonationForms/Blocks/DonationFormBlock/resources/app/index.tsx
@@ -1,9 +1,9 @@
-import {createRoot} from '@wordpress/element';
 import {__} from '@wordpress/i18n';
 import IframeResizer from 'iframe-resizer-react';
 import isRouteInlineRedirect from '@givewp/forms/app/utilities/isRouteInlineRedirect';
 import ModalForm from '@givewp/src/Campaigns/Blocks/shared/components/ModalForm';
 import '../editor/styles/index.scss';
+import renderDonationForm from './renderDonationForm';
 
 /**
  * @since 3.2.1 Revert the display style value of "fullForm" to "onpage".
@@ -94,32 +94,9 @@ export default function DonationFormBlockApp({
 const roots = document.querySelectorAll('.root-data-givewp-embed');
 
 /**
+ * @unreleased update to use renderDonationForm
  * @since 3.22.0 Add locale support
  */
-roots.forEach((root) => {
-    let dataSrcUrl = root.getAttribute('data-src');
-    const locale = root.getAttribute('data-form-locale');
-    if (locale) {
-        const url = new URL(dataSrcUrl);
-        url.searchParams.set('locale', locale);
-        dataSrcUrl = url.toString();
-    }
-
-    const dataSrc = dataSrcUrl;
-    const embedId = root.getAttribute('data-givewp-embed-id');
-    const formFormat = root.getAttribute('data-form-format');
-    const openFormButton = root.getAttribute('data-open-form-button');
-    const formUrl = root.getAttribute('data-form-url');
-    const formViewUrl = root.getAttribute('data-form-view-url');
-
-    createRoot(root).render(
-        <DonationFormBlockApp
-            openFormButton={openFormButton}
-            formFormat={formFormat}
-            dataSrc={dataSrc}
-            embedId={embedId}
-            formUrl={formUrl}
-            formViewUrl={formViewUrl}
-        />
-    );
+roots?.forEach((root) => {
+    renderDonationForm(root);
 });
