+++ src/ThirdPartySupport/Elementor/Widgets/V2/ElementorDonationFormWidget/resources/widget.jsx
@@ -1,5 +1,4 @@
-import {createRoot} from '@wordpress/element';
-import DonationFormBlockApp from '@givewp/src/DonationForms/Blocks/DonationFormBlock/resources/app';
+import renderDonationForm from '@givewp/src/DonationForms/Blocks/DonationFormBlock/resources/app/renderDonationForm';
 
 import './widget.scss';
 
@@ -7,32 +6,8 @@ export default class DonationFormWidget extends elementorModules.frontend.handle
     render() {
         const roots = document.querySelectorAll(`[data-id="${this.getID()}"] .root-data-givewp-embed`);
 
-        roots.forEach((root) => {
-            let dataSrcUrl = root.getAttribute('data-src');
-            const locale = root.getAttribute('data-form-locale');
-            if (locale) {
-                const url = new URL(dataSrcUrl);
-                url.searchParams.set('locale', locale);
-                dataSrcUrl = url.toString();
-            }
-
-            const dataSrc = dataSrcUrl;
-            const embedId = root.getAttribute('data-givewp-embed-id');
-            const formFormat = root.getAttribute('data-form-format');
-            const openFormButton = root.getAttribute('data-open-form-button');
-            const formUrl = root.getAttribute('data-form-url');
-            const formViewUrl = root.getAttribute('data-form-view-url');
-
-            createRoot(root).render(
-                <DonationFormBlockApp
-                    openFormButton={openFormButton}
-                    formFormat={formFormat}
-                    dataSrc={dataSrc}
-                    embedId={embedId}
-                    formUrl={formUrl}
-                    formViewUrl={formViewUrl}
-                />
-            );
+        roots?.forEach((root) => {
+            renderDonationForm(root);
         });
     }
